Base de datos/proyectos/crud_docker/requirements.txt
streamlit
streamlit-option-menu
python-dotenv
mysql-connector-python
psycopg2
pandas